+++ python_module/src/session_getter_setter.cpp
@@ -489,6 +489,12 @@ const VectorMapND & session::get_out_total_init_maxgate_readout_times() const { 
 
 const VectorMapND & session::get_out_z_op_expects() const { return session::out_z_op_expects_ ; }
 //
+
+const std::shared_ptr<xacc::ExecutionInfo::WaveFuncType> & session::get_state_vec_raw() const {return session::state_vec_ ;}
+//
+
+void session::get_state_vec(const bool &in_get_state_vec) {session::in_get_state_vec_ = in_get_state_vec ;}
+//
 void session::set_noise_mitigation(const std::string &noise_mitigation) {
   session::validate_noise_mitigation(noise_mitigation);
   session::error_mitigations_.clear();